requirements.txt
absl-py==0.14.1
astunparse==1.6.3
atari-py==0.2.9
backcall==0.2.0
box2d-py==2.3.8
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.2
cloudpickle==1.2.2
colorama==0.4.4
cycler==0.10.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
decorator==5.1.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.5.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gym==0.10.5
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
Markdown==3.3.4
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
# Editable install with no version control (multiagent==0.0.1)
-e c:\codelibs\multiagent-particle-envs
nltk==3.6.3
numpy==1.19.5
numpy-stl==2.16.3
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.2
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1625663293114/work
prompt-toolkit==3.0.20
protobuf==3.18.1
ptan==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.1.9
pyglet==1.5.0
Pygments==2.10.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-utils==2.5.6
pytorch-ignite==0.4.7
pytz==2021.3
PyYAML==6.0
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scipy==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
torch==1.9.1
torchaudio==0.9.1
torchvision==0.2.2
tornado @ file:///C:/ci/tornado_1606942379977/work
tqdm==4.62.3
traitlets==4.3.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1632313171031/work
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
